Catch StompException thrown when closing socket
If Ctrl+C is pressed to quit the worker, SlmQueue catches the SIG*
commands nicely but the socket is disconnected and Stomp PHP throws
a nasty exception. Kind of dirty fix this by catching the specific
exception (which is used in multiple places) and only silently
ignore the exception when the message equals the string we are looking
for.