Subscribe only when not subscribed yet
Use the state stored in the stomp client to check if a subscription
has been made already. This only works:

* If requests are send asynchronously (default true -- sync = false)
* If a sync request returns a reciept with corect id

Because of this, the only way this fails if a subscription is made
synchronously and the reciept hasn't been received yet but either
way a second subscription is already made. This is quite an unlikely
scenario so we can -for now- safely assume this is the best way.